+++ common/tools.h
@@ -2,4 +2,4 @@
 #include <stdint.h>
 
 void reverse(char *str, int length);
-void itoa7(int16_t num, char *str);
+void itoa7(int32_t num, char *str);
